+++ src/common/init.c
@@ -37,6 +37,7 @@ static void defaultConfig(void)
     ffStrbufInit(&instance.config.colorKeys);
     ffStrbufInit(&instance.config.colorTitle);
     ffStrbufInitS(&instance.config.keyValueSeparator, ": ");
+    instance.config.processingTimeout = 1000;
 
     #if defined(__linux__) || defined(__FreeBSD__)
     ffStrbufInit(&instance.config.playerName);