+++ src/common/processing.h
@@ -5,8 +5,6 @@
 
 #include "util/FFstrbuf.h"
 
-#define FF_WAIT_TIMEOUT 1000
-
 const char* ffProcessAppendOutput(FFstrbuf* buffer, char* const argv[], bool useStdErr);
 
 static inline const char* ffProcessAppendStdOut(FFstrbuf* buffer, char* const argv[])