+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Features:
 * Add `--logo-type small` to search for small logos
 * Support detecting brightness of external displays with DDC/CI (guard behind `--allow-slow-operations`) (Brightness)
 * Add option `--size-ndigits` and `--size-max-prefix` (#494)
+* Add option `--processing-timeout` to the timeout when waiting for child processes.
 
 # 1.12.2